requirements.txt
pandas>=2.0
openpyxl>=3.1
requests>=2.31
httpx>=0.27
beautifulsoup4>=4.12
pypdf>=4.0
pdfplumber>=0.11
tenacity>=8.2
python-slugify>=8.0
rapidfuzz>=3.6
tldextract>=5.1
waybackpy>=3.0
